LPS-82509 SF, move static methods into one class
