+++ modules/apps/portal/portal-cluster-multiple/src/main/java/com/liferay/portal/cluster/multiple/internal/io/ClusterSerializerUtil.java
@@ -17,17 +17,21 @@ package com.liferay.portal.cluster.multiple.internal.io;
 import com.liferay.petra.io.ProtectedObjectInputStream;
 import com.liferay.petra.io.SerializationConstants;
 import com.liferay.petra.io.unsync.UnsyncByteArrayInputStream;
+import com.liferay.petra.io.unsync.UnsyncByteArrayOutputStream;
 import com.liferay.petra.lang.ClassResolverUtil;
 
 import java.io.IOException;
 import java.io.InputStream;
 import java.io.ObjectInputStream;
+import java.io.ObjectOutputStream;
 import java.io.ObjectStreamClass;
+import java.io.OutputStream;
+import java.io.Serializable;
 
 /**
- * @author Lance Ji
+ * @author Tina Tian
  */
-public class ClusterDeserializer {
+public class ClusterSerializerUtil {
 
 	public static Object readObject(byte[] bytes, int offset, int length)
 		throws ClassNotFoundException {
@@ -50,6 +54,51 @@ public class ClusterDeserializer {
 		catch (IOException ioe) {
 			throw new RuntimeException(ioe);
 		}
+	}
+
+	public static byte[] writeObject(Serializable serializable) {
+		try {
+			UnsyncByteArrayOutputStream unsyncByteArrayOutputStream =
+				new UnsyncByteArrayOutputStream();
+
+			unsyncByteArrayOutputStream.write(SerializationConstants.TC_OBJECT);
+
+			ObjectOutputStream objectOutputStream =
+				new ClusterAnnotatedObjectOutputStream(
+					unsyncByteArrayOutputStream);
+
+			objectOutputStream.writeObject(serializable);
+
+			objectOutputStream.flush();
+
+			return unsyncByteArrayOutputStream.unsafeGetByteArray();
+		}
+		catch (IOException ioe) {
+			throw new RuntimeException(
+				"Unable to write ordinary serializable object " + serializable,
+				ioe);
+		}
+	}
+
+	private static class ClusterAnnotatedObjectOutputStream
+		extends ObjectOutputStream {
+
+		@Override
+		protected void annotateClass(Class<?> clazz) throws IOException {
+			ClassLoader classLoader = clazz.getClassLoader();
+
+			String contextName = ClusterClassLoaderPool.getContextName(
+				classLoader);
+
+			writeUTF(contextName);
+		}
+
+		private ClusterAnnotatedObjectOutputStream(OutputStream outputStream)
+			throws IOException {
+
+			super(outputStream);
+		}
+
 	}
 
 	private static class ClusterProtectedAnnotatedObjectInputStream
